analytics-service/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.28.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-json-logger==2.0.7
redis==5.0.1
pandas==2.1.4
numpy==1.24.3
asyncio-mqtt==0.16.1
websockets==12.0
python-multipart==0.0.6
